+++ src/test/java/lotto/LottoResultTest.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.domain.LottoResult;
+import lotto.domain.MatchStatus;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
